+++ src/trx/game/cutscene.c
@@ -292,7 +292,7 @@ static void M_PlayerControl(const int16_t item_num)
 
     int16_t floor_room_num = item->room_num;
     const SECTOR *const sector = Room_GetSector(pos, &floor_room_num);
-    const int16_t height = Room_GetHeight(sector, pos);
+    const int32_t height = Room_GetHeight(sector, pos);
     item->floor = height == NO_HEIGHT ? pos.y : height;
 
     Lara_Animate(item);
