+++ src/trx/game/objects/general/cutscene_player.c
@@ -33,7 +33,7 @@ static void M_Control(const int16_t item_num)
 
     int16_t floor_room_num = item->room_num;
     const SECTOR *const sector = Room_GetSector(pos, &floor_room_num);
-    const int16_t height = Room_GetHeight(sector, pos);
+    const int32_t height = Room_GetHeight(sector, pos);
     item->floor = height == NO_HEIGHT ? pos.y : height;
 
     if (item->dynamic_light && item->status != IS_INVISIBLE) {
